requirements.txt
docling
faiss-cpu
sentence-transformers
transformers
torch 
requests
numpy
llama-cpp-python[server] 
uvicorn
requests
nomic